use ApplicativePlus as Alternative ... fail
