+++ core/src/main/scala/scalaz/Applicative.scala
@@ -17,7 +17,7 @@ package scalaz
  *  @see [[scalaz.Applicative.ApplicativeLaw]]
  */
 ////
-trait Applicative[F[_]] extends Apply[F] with ApplicativeDivisible[F] { self =>
+trait Applicative[F[_]] extends Apply[F] { self =>
   ////
   def point[A](a: => A): F[A]
 
@@ -30,9 +30,6 @@ trait Applicative[F[_]] extends Apply[F] with ApplicativeDivisible[F] { self =>
 
   override def apply2[A, B, C](fa: => F[A], fb: => F[B])(f: (A, B) => C): F[C] =
     ap2(fa, fb)(point(f))
-
-  // ApplicativeDivisible impl
-  override def xproduct0[Z](z: =>Z): F[Z] = pure(z)
 
   // impls of sequence, traverse, etc
 
@@ -131,7 +128,7 @@ object Applicative {
   ////
 }
 
-trait IsomorphismApplicative[F[_], G[_]] extends Applicative[F] with IsomorphismApply[F, G] with IsomorphismApplicativeDivisible[F, G]{
+trait IsomorphismApplicative[F[_], G[_]] extends Applicative[F] with IsomorphismApply[F, G]{
   implicit def G: Applicative[G]
   ////
