+++ core/src/main/scala/scalaz/Decidable.scala
@@ -10,7 +10,7 @@ package scalaz
  * https://hackage.haskell.org/package/contravariant-1.4.1/docs/Data-Functor-Contravariant-Divisible.html#t:Decidable
  */
 ////
-trait Decidable[F[_]] extends Divisible[F] with CoapplicativeDecidable[F] { self =>
+trait Decidable[F[_]] extends Divisible[F] with Derives[F] { self =>
   ////
 
   final def choose[Z, A1, A2](a1: =>F[A1], a2: =>F[A2])(f: Z => A1 \/ A2): F[Z] = choose2(a1, a2)(f)
@@ -45,6 +45,21 @@ trait Decidable[F[_]] extends Divisible[F] with CoapplicativeDecidable[F] { self
   )(implicit fa1: F[A1], fa2: F[A2], fa3: F[A3], fa4: F[A4]): F[Z] =
     choose4(fa1, fa2, fa3, fa4)(f)
   // ... choosingX
+
+  override def xproduct0[Z](z: =>Z): F[Z] = conquer
+  override def xproduct1[Z, A1](a1: F[A1])(f: A1 => Z, g: Z => A1): F[Z] = xmap(a1, f, g)
+  override def xproduct2[Z, A1, A2](a1: =>F[A1], a2: =>F[A2])(
+    f: (A1, A2) => Z,
+    g: Z => (A1, A2)
+  ): F[Z] = divide2(a1, a2)(g)
+  override def xproduct3[Z, A1, A2, A3](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3])(
+    f: (A1, A2, A3) => Z,
+    g: Z => (A1, A2, A3)
+  ): F[Z] = divide3(a1, a2, a3)(g)
+  override def xproduct4[Z, A1, A2, A3, A4](a1: =>F[A1], a2: =>F[A2], a3: =>F[A3], a4: =>F[A4])(
+    f: (A1, A2, A3, A4) => Z,
+    g: Z => (A1, A2, A3, A4)
+  ): F[Z] = divide4(a1, a2, a3, a4)(g)
 
   override def xcoproduct1[Z, A1](a1: =>F[A1])(
     f: A1 => Z,
@@ -93,7 +108,7 @@ object Decidable {
   ////
 }
 
-trait IsomorphismDecidable[F[_], G[_]] extends Decidable[F] with IsomorphismDivisible[F, G] with IsomorphismCoapplicativeDecidable[F, G]{
+trait IsomorphismDecidable[F[_], G[_]] extends Decidable[F] with IsomorphismDivisible[F, G] with IsomorphismDerives[F, G]{
   implicit def G: Decidable[G]
   ////
